Cast SPI_processed to C.uint64 (maybe solves #3)

+++ pl.go
@@ -576,7 +576,7 @@ func (stmt *Stmt) Query(args ...interface{}) (*Rows, error) {
 	valuesP, nullsP := spiArgs(args)
 	rv := C.SPI_execute_plan(stmt.spiPlan, valuesP, nullsP, C.true, 0)
 	if rv == C.SPI_OK_SELECT && C.SPI_processed > 0 {
-		return newRows(C.SPI_tuptable.vals, C.SPI_tuptable.tupdesc, C.SPI_processed), nil
+		return newRows(C.SPI_tuptable.vals, C.SPI_tuptable.tupdesc, C.uint64(C.SPI_processed)), nil
 	}
 	return nil, fmt.Errorf("Query failed: %s", C.GoString(C.SPI_result_code_string(C.SPI_result)))
 }
